requirements.txt
async-generator==1.10
asyncio==3.4.3
attrs==22.1.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.5
commonmark==0.9.1
defusedxml==0.7.1
dj-database-url==0.5.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==22.0.0
h11==0.13.0
httpie==3.2.1
jmespath==0.9.5
joblib==1.3.2
MarkupSafe==2.1.5
match==0.3.2
multidict==6.0.2
mysql-connector-python==8.0.32
nltk==3.8.1
outcome==1.2.0
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.9.4
Pygments==2.13.0
PySocks==1.7.1
pytz==2019.3
regex==2023.12.25
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
robotframework==5.0.1
robotframework-pythonlibcore==3.0.0
robotframework-seleniumlibrary==6.0.0
selenium==4.4.3
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.66.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.25.11
websocket==0.2.1
websockets==10.4
whitenoise==5.0.1
wsproto==1.1.0
zope.event==4.6
zope.interface==5.5.2
